pyproject.toml
[project]
name = "digital-cto"
version = "0.1.0"
description = "AfCEN Digital CTO - AI-powered multi-agent technical leadership system"
requires-python = ">=3.9"
dependencies = [
    # Agent orchestration
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",

    # API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",

    # Memory stores
    "redis[hiredis]>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "qdrant-client>=1.12.0",
    "alembic>=1.14.0",

    # GitHub integration
    "PyGithub>=2.5.0",
    "httpx>=0.28.0",

    # WebSocket for OpenClaw/JARVIS
    "websockets>=14.0",

    # Scheduler
    "apscheduler>=3.10.0",

    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",

    # Production: Retry & Resilience
    "tenacity>=8.5.0",

    # Production: Rate Limiting
    "slowapi>=0.1.9",

    # Production: Structured Logging
    "python-json-logger>=2.0.7",

    # Production: Metrics
    "prometheus-client>=0.21.0",

    # Security scanning
    "semgrep>=1.100.0",

    # Phase 4: Docker
    "docker>=7.0.0",

    # TUI: Terminal User Interface
    "click>=8.1.0",
    "questionary>=2.0.0",
    "rich>=13.0.0",
]

[project.scripts]
cto = "src.tui.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
